Refacotry gang management code.
This commit is a first step to abstract gang management and dispatch code in
order to allow different implementation easier.
