+++ src/include/cdb/cdbgangmgr.h
@@ -1,0 +1,18 @@
+/*-------------------------------------------------------------------------
+ *
+ * cdbgangmgr.h
+ *
+ * Copyright (c) 2005-2016, Pivotal Inc.
+ *
+ *-------------------------------------------------------------------------
+ */
+
+#ifndef _CDBGANGMGR_H_
+#define _CDBGANGMGR_H_
+
+#include "postgres.h"
+#include "cdb/cdbgang.h"
+
+GangMgrMethods GetGangMgr();
+
+#endif   /* _CDBGANGMGR_H_ */
